........
  r19284 | pablo.kraan | 2010-09-01 13:16:32 -0300 (Wed, 01 Sep 2010) | 1 line
  
  Adds flags to disable the purge of queues on AbstractJmsFunctionalTestCase. Change required for new tests of EE-1999
........


git-svn-id: https://svn.codehaus.org/mule/branches/mule-3.x@20340 bf997673-6b11-0410-b953-e057580c5b09

+++ transports/jms/src/test/java/org/mule/transport/jms/integration/AbstractJmsFunctionalTestCase.java
@@ -118,6 +118,8 @@ public abstract class AbstractJmsFunctionalTestCase extends FunctionalTestCase
     protected Scenario scenarioRollback;
     protected Scenario scenarioNotReceive;
     protected Scenario scenarioReceive;
+    protected boolean purgeQueuesOnPreSetUp = true;
+    protected boolean purgeQueuesOnTearDown = true;
 
     private MuleClient client = null;
 
@@ -222,9 +224,12 @@ public abstract class AbstractJmsFunctionalTestCase extends FunctionalTestCase
     {
         super.suitePreSetUp();
 
-        purge(getInboundQueueName());
-        purge(getOutboundQueueName());
-        // TODO DZ: get all of the queue/topic names from the Mule config and just purge those 
+        if (purgeQueuesOnPreSetUp)
+        {
+            purge(getInboundQueueName());
+            purge(getOutboundQueueName());
+            // TODO DZ: get all of the queue/topic names from the Mule config and just purge those
+        }
     }
 
     /**
@@ -527,9 +532,12 @@ public abstract class AbstractJmsFunctionalTestCase extends FunctionalTestCase
 
     protected void doTearDown() throws Exception
     {
-        purge(getInboundQueueName());
-        purge(getOutboundQueueName());
-        purgeTopics();
+        if (purgeQueuesOnTearDown)
+        {
+            purge(getInboundQueueName());
+            purge(getOutboundQueueName());
+            purgeTopics();
+        }
 
         super.doTearDown();
         if (client != null)
